Add files via upload
This is a prototype of a chess board game. It is not optimized. This allows a player to make any valid move and play chess, however it does not check for a check/checkmate. Nor does it end, it is up to the user to determine this. There are a lot of problems with this such as not freeing memory when a piece is taken, messy code, mispelled file names, etc. That being said this was an initial attempt so that I could first understand some of the minor things about chess that I would need to know before starting. I plan on making a new project with a GUI and that is where this project will continue